Changes from branch merged. Added S9 support. Added fix for MF.

+++ src/com/almalence/opencam/ui/AlmalenceGUI.java
@@ -2514,8 +2514,8 @@ public class AlmalenceGUI extends GUI implements SeekBar.OnSeekBarChangeListener
 					{
 						float initValue = preferences.getFloat(MainScreen.sFocusDistancePref,
 								CameraController.getMinimumFocusDistance());
-						focusBar.setMax((int) CameraController.getMinimumFocusDistance() * 100);
-						focusBar.setProgress((int) initValue * 100);
+						focusBar.setMax((int) (CameraController.getMinimumFocusDistance() * 100));
+						focusBar.setProgress((int) (initValue * 100));
 
 						TextView leftText = (TextView) guiView.findViewById(R.id.focusDistanceLeftText);
 						TextView rightText = (TextView) guiView.findViewById(R.id.focusDistanceRightText);
@@ -7881,9 +7881,9 @@ public class AlmalenceGUI extends GUI implements SeekBar.OnSeekBarChangeListener
 		} else if (seekBar == (SeekBar) guiView.findViewById(R.id.focusDistanceSeekBar))
 		{
 			int iDistance = progress;
-			CameraController.setCameraFocusDistance(iDistance / 100);
-			preferences.edit().putFloat(MainScreen.sFocusDistancePref, (float) iDistance / 100).commit();
-			mFocusDistance = iDistance / 100;
+			CameraController.setCameraFocusDistance(((float)iDistance) / 100);
+			preferences.edit().putFloat(MainScreen.sFocusDistancePref, ((float) iDistance) / 100).commit();
+			mFocusDistance = ((float) iDistance) / 100;
 		} else if (seekBar == (SeekBar) guiView.findViewById(R.id.manualWBSeekBar))
 		{
 			int iReadableTemp = (progress + 10) * 100;
